requirements.txt
dash==3.0.4
numpy==1.26.4
pandas==2.1.4
dash-mantine-components==1.2.0
dash-ag-grid==31.3.1
python-dotenv==1.0.1
dash-iconify==0.1.2
dash-bootstrap-components==2.0.2
pyspark==3.5.1
pytz==2024.1
bandit==1.8.6
ruff==0.12.2